feat: made kitty clicker work like described in course
... okay, so I mostly "referenced" it from the course materials
https://github.com/udacity/ud989-cat-clicker-premium-vanilla/blob/master/js/app.js
but it wasn't making sense any other way and I really just needed
to get it done. Even in doing that I had to debug it the hard way
because it was still throwing errors at me, mostly to do with the
renaming of various variables from "cat" to "kitty" to make it make
more sense to me.